Add class-based interface
